create parent class
add a constructor nethos with one parameter, include values for Media. Add getters that return the value saved to the matching property. Include setter and a method that changes the value saved to the _isCheckedOut pro. Add a method getAverageRating that returns the ave value of the ratings array.  Then add a method addRating with .push() to add it to the end of the ratings array.