Create HighSchool class
Create an empty HighSchool class that extends School.
Create a constructor() that accepts arguments for name, numberOfStudents, and sportsTeams.
On the first line of the constructor(), call super and pass it name, 'high', and numberOfStudents as arguments.
On the next line, set the sportsTeams property using this._
Under the constructor(), create a getter for sportsTeams. Use console.log() to print the sportsTeams value to the console.